+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 Added getter and setter for Random object in AbstractOperator. Creates tests for all abstract classes. Mutation and
 selection operators now throw exceptions if illegal argument is provided in constructor or setter.
 AbstractSeectionOperator now specifies a single abstract method that needs to be overridden in order to perform
-selection. AbstractPopulationGenerator now throws exception if requested population size is less than 0.
+selection. AbstractPopulationGenerator now throws exception if requested population size is less than 0. Corrected
+comparison method in AbstractChromosome.
 
 ### 0.2.1